refactor(ui): les statistiques de Guyane sont en tsx
